fix: slim down the tests
